Time: 39 ms (8.43%), Space: 48.5 MB (5.02%) - LeetHub

+++ 0322-coin-change/0322-coin-change.java
@@ -1,20 +1,20 @@
 class Solution {
-    public long Chck(int i , int[] coins ,int amount , int[][] dp){
+    public long Chck(int i , int[] coins ,int amount , long[][] dp){
         if(i == coins.length){
             if(0 == amount) return 0;
             return Integer.MAX_VALUE;
         }
         if(dp[i][amount] != -1) return dp[i][amount];
         long skip = Chck(i+1 , coins , amount , dp);
-        if(amount-coins[i] < 0) return dp[i][amount] = (int)skip;
+        if(amount-coins[i] < 0) return dp[i][amount] = skip;
         long pick = 1 + Chck(i , coins , amount-coins[i] , dp);  
-        return  dp[i][amount] = (int) Math.min(pick , skip);
+        return  dp[i][amount] = Math.min(pick , skip);
     }
     public int coinChange(int[] coins, int amount) {
         int n = coins.length; 
-        int[][] dp = new int[n][amount+1];
-        for(int[] row : dp) Arrays.fill(row,-1);
-        dp[n-1][amount] = (int) Chck(0,coins,amount,dp);
-        return (dp[n-1][amount] == Integer.MAX_VALUE)? -1 : dp[n-1][amount];
+        long[][] dp = new long[n][amount+1];
+        for(long[] row : dp) Arrays.fill(row,-1);
+        dp[n-1][amount] = Chck(0,coins,amount,dp);
+        return (dp[n-1][amount] == Integer.MAX_VALUE)? -1 : (int) dp[n-1][amount];
     }
 }
